Clean up license copy-paste bugs from intel-tex-rs-2
The licenses were copied form another repo that had the wrong author in
`LICENSE-MIT`, but was not noticed by me in `git status` after only
seeing "rename LICENSE to LICENSE-MIT".

Furthermore the appendix in `LICENSE-APACHE` is not needed, this is only
a hint as to what'd need to be copied into every file file header if we
were to have per-file headers, which we don't [1] - and the indentation
has been fixed so that the file is now identical to the one used by
rust-lang themselves.

[1]: https://github.com/rust-lang/rust/commit/2ccf65c7eba519dcc407cb76f08d04ac00d05851