Upgrade to `ispc 1.1.0` with cleaned up bindings file
`ispc_compile` can now be configured to omit irrelevant bindings from
the bindings file and leave only the function(s) in place that are
relevant to interface with the ISPC kernel
(https://github.com/Twinklebear/ispc-rs/pull/20).  This makes the file
less busy/noisy, and it won't change when regenerating across OS-es
anymore.