lanczos3: Scale input kernel by source size, not target size (#26)
When the 7x7 kernel runs over pixels in the _source_ image, its offset
is multiplied by the inverse of the _target_ image size to turn these
integer coordinates into float coordinates (only to turn them back into
integer on the next line :facepalm:). This means that the floating-point
offset of the kernel relative to the center of the source pixel we are
sampling is now relative to the dimensions of the target image instead
of the source image, leading us to skip source pixels as the target is
smaller than the source and thus the offset is larger than one step in
source pixels.